+++ api/src/main/java/com/tersesystems/echopraxia/api/Value.java
@@ -1,5 +1,7 @@
 package com.tersesystems.echopraxia.api;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.*;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -46,20 +48,20 @@ public abstract class Value<V> {
   public String toString() {
     final Object raw = raw();
     final Type type = type();
-    if (raw == null
-        || type == Type.NULL) { // if null value or a raw value was set to null, keep going.
+    if (raw == null || type == Type.NULL) { // if null value or a raw value was set to null, keep going.
       return "null";
     }
+
     if (type == Type.STRING) {
-      return ((String) raw());
+      return ((String) raw);
     }
 
     if (type == Type.BOOLEAN) {
-      return ((Boolean) raw()) ? Boolean.TRUE.toString() : Boolean.FALSE.toString();
+      return Boolean.toString((Boolean) raw);
     }
 
     if (type == Type.NUMBER) {
-      return raw().toString();
+      return raw.toString();
     }
 
     final StringBuilder b = new StringBuilder(255);
@@ -78,6 +80,22 @@ public abstract class Value<V> {
     return new StringValue(value);
   }
 
+  @NotNull
+  public static Value<Byte> number(@NotNull Byte value) {
+    final int offset = 128;
+    return NumberValue.ByteValue.Cache.cache[(int) value + offset];
+  }
+
+  @NotNull
+  public static Value<Short> number(@NotNull Short s) {
+    final int offset = 128;
+    int sAsInt = s;
+    if (sAsInt >= -128 && sAsInt <= 127) { // must cache
+      return NumberValue.ShortValue.Cache.cache[sAsInt + offset];
+    }
+    return new NumberValue.ShortValue(s);
+  }
+
   /**
    * Wraps a number with a Value.
    *
@@ -85,8 +103,42 @@ public abstract class Value<V> {
    * @return the Value
    */
   @NotNull
-  public static NumberValue number(@NotNull Number value) {
-    return new NumberValue(value);
+  public static NumberValue<Integer> number(@NotNull Integer value) {
+    if (value >= NumberValue.IntegerValue.Cache.low && value <= NumberValue.IntegerValue.Cache.high)
+      return NumberValue.IntegerValue.Cache.cache[value + (-NumberValue.IntegerValue.Cache.low)];
+    return new NumberValue.IntegerValue(value);
+  }
+
+  public static NumberValue<Long> number(@NotNull Long value) {
+    final int offset = 128;
+    if (value >= -128 && value <= 127) { // will cache
+      return NumberValue.LongValue.Cache.cache[value.intValue() + offset];
+    }
+    return new NumberValue.LongValue(value);
+  }
+
+  public static NumberValue<Float> number(@NotNull Float value) {
+    if (Objects.equals(value, 0.0f)) return NumberValue.FloatValue.ZERO;
+    return new NumberValue.FloatValue(value);
+  }
+
+  public static NumberValue<Double> number(@NotNull Double value) {
+    if (Objects.equals(value, 0.0d)) return NumberValue.DoubleValue.ZERO;
+    return new NumberValue.DoubleValue(value);
+  }
+
+  public static NumberValue<BigInteger> number(@NotNull BigInteger value) {
+    if (Objects.equals(value, BigInteger.ZERO)) return NumberValue.BigIntegerValue.ZERO;
+    return new NumberValue.BigIntegerValue(value);
+  }
+
+  public static NumberValue<BigDecimal> number(@NotNull BigDecimal value) {
+    if (Objects.equals(value, BigDecimal.ZERO)) return NumberValue.BigDecimalValue.ZERO;
+    return new NumberValue.BigDecimalValue(value);
+  }
+
+  public static <N extends Number & Comparable<N>> NumberValue<N> number(@NotNull N value) {
+    return new NumberValue<>(value);
   }
 
   /**
@@ -153,7 +205,7 @@ public abstract class Value<V> {
   /**
    * Wraps an array of values with string values.
    *
-   * @param values varadic elements of values.
+   * @param values variadic elements of values.
    * @return the Value.
    */
   @NotNull
@@ -167,11 +219,12 @@ public abstract class Value<V> {
   /**
    * Wraps an array of values with number values.
    *
-   * @param values varadic elements of values.
+   * @param values variadic elements of values.
    * @return the Value.
    */
+  @SafeVarargs
   @NotNull
-  public static ArrayValue array(Number @NotNull ... values) {
+  public static <N extends Number & Comparable<N>> ArrayValue array(N @NotNull ... values) {
     if (values.length == 0) {
       return ArrayValue.EMPTY;
     }
@@ -380,15 +433,16 @@ public abstract class Value<V> {
     }
   }
 
-  public static final class NumberValue extends Value<Number> {
-    private final Number raw;
+  public static class NumberValue<N extends Number & Comparable<N>> extends Value<N>
+      implements Comparable<NumberValue<N>> {
+    private final N raw;
 
-    private NumberValue(Number number) {
+    private NumberValue(N number) {
       this.raw = number;
     }
 
     @Override
-    public Number raw() {
+    public N raw() {
       return raw;
     }
 
@@ -401,7 +455,7 @@ public abstract class Value<V> {
     public boolean equals(Object o) {
       if (this == o) return true;
       if (o == null || getClass() != o.getClass()) return false;
-      NumberValue that = (NumberValue) o;
+      NumberValue<?> that = (NumberValue<?>) o;
       // follow the Java example of not allowing general comparison here
       // integer != long, long != float, float != double etc.
       // if you want specific comparison, then cast the raw value.
@@ -412,9 +466,114 @@ public abstract class Value<V> {
     public int hashCode() {
       return raw != null ? raw.hashCode() : 0;
     }
+
+    @Override
+    public int compareTo(@NotNull Value.NumberValue<N> o) {
+      return this.raw.compareTo(o.raw);
+    }
+
+    private static final class ByteValue extends NumberValue<Byte> {
+      private ByteValue(Byte number) {
+        super(number);
+      }
+
+      private static class Cache {
+        private Cache() {}
+
+        static final ByteValue[] cache = new ByteValue[-(-128) + 127 + 1];
+
+        static {
+          for (int i = 0; i < cache.length; i++) cache[i] = new ByteValue((byte) (i - 128));
+        }
+      }
+    }
+
+    private static final class ShortValue extends NumberValue<Short> {
+      public static final ShortValue ZERO = new ShortValue((short) 0);
+
+      private ShortValue(Short number) {
+        super(number);
+      }
+
+      private static class Cache {
+        private Cache() {}
+
+        static final ShortValue[] cache = new ShortValue[-(-128) + 127 + 1];
+
+        static {
+          for (int i = 0; i < cache.length; i++) cache[i] = new ShortValue((short) (i - 128));
+        }
+      }
+    }
+
+    private static final class IntegerValue extends NumberValue<Integer> {
+      private IntegerValue(Integer number) {
+        super(number);
+      }
+
+      private static class Cache {
+        static final int low = -128;
+        static final int high = 127;
+
+        static final IntegerValue[] cache = new IntegerValue[(high - low) + 1];
+
+        static {
+          for (int i = 0; i < cache.length; i++) cache[i] = new IntegerValue((i - 128));
+        }
+        private Cache() {}
+      }
+    }
+
+    private static final class LongValue extends NumberValue<Long> {
+      private LongValue(Long number) {
+        super(number);
+      }
+
+      private static class Cache {
+        private Cache() {}
+
+        static final LongValue[] cache = new LongValue[-(-128) + 127 + 1];
+
+        static {
+          for (int i = 0; i < cache.length; i++) cache[i] = new LongValue((long) (i - 128));
+        }
+      }
+    }
+
+    private static final class DoubleValue extends NumberValue<Double> {
+      public static final DoubleValue ZERO = new DoubleValue(0.0d);
+
+      private DoubleValue(Double number) {
+        super(number);
+      }
+    }
+
+    private static final class FloatValue extends NumberValue<Float> {
+      public static final FloatValue ZERO = new FloatValue(0.0f);
+
+      private FloatValue(Float number) {
+        super(number);
+      }
+    }
+
+    private static final class BigIntegerValue extends NumberValue<BigInteger> {
+      public static final BigIntegerValue ZERO = new BigIntegerValue(java.math.BigInteger.ZERO);
+
+      private BigIntegerValue(BigInteger number) {
+        super(number);
+      }
+    }
+
+    private static final class BigDecimalValue extends NumberValue<BigDecimal> {
+      public static final BigDecimalValue ZERO = new BigDecimalValue(java.math.BigDecimal.ZERO);
+
+      private BigDecimalValue(BigDecimal number) {
+        super(number);
+      }
+    }
   }
 
-  public static final class StringValue extends Value<String> {
+  public static final class StringValue extends Value<String> implements Comparable<StringValue> {
     private final String raw;
 
     private StringValue(String s) {
@@ -442,6 +601,11 @@ public abstract class Value<V> {
     @Override
     public int hashCode() {
       return raw != null ? raw.hashCode() : 0;
+    }
+
+    @Override
+    public int compareTo(@NotNull Value.StringValue o) {
+      return o.raw.compareTo(this.raw);
     }
   }